refactoring frontend calls into service to facilitate testing; refactored disconnect
